Project can now be worked with in CLion.
Added a simple CMake file and some more complete paths for a few headers in includes.

Cmake can build the project itself or (better) delegate make to do it.

Tests are still to be run from terminal. (or put run-test.sh as executable in CLion's build configuration to be run after each new build).

CLion's terminal emulator does not allow the terminal to run in curses mode. Still broken in 1.5 EAP.